test(ui): hoist shared jsdom shims and withTooltip helper into src/test (Closes #1174)
The same test boilerplate was duplicated across ~35 test files:

- An idempotent jsdom shim block (ResizeObserver + pointer-capture, plus the
  Radix Select scrollIntoView shim) copy-pasted into 20+ files. These are
  global, environment-level, and idempotent, so they now live once in the
  central src/test/setup.ts (already wired as setupFiles). The DOM-only shims
  are guarded behind a `typeof Element !== "undefined"` check so the shared
  setup stays safe in the node-environment bridge tests.
- A byte-identical `withTooltip(ui) => <TooltipProvider delayDuration={0}>`
  helper redefined in 16 files, now a single shared src/test/tooltip.tsx module
  adopted everywhere (including the createElement-based inline wraps).

Behavior is unchanged; no test assertion was altered. The Tooltip primitive's
own test keeps its local provider wrapper since it exercises the primitive
directly.
