+++ src/components/EmbeddedServerSidebar/EmbeddedServerDialog.test.tsx
@@ -42,16 +42,6 @@ describe("EmbeddedServerDialog", () => {
     container = document.createElement("div");
     document.body.appendChild(container);
     root = createRoot(container);
-    // Radix Select (the migrated bind-address dropdown) probes pointer-capture
-    // and scroll APIs that jsdom omits.
-    if (!Element.prototype.hasPointerCapture) {
-      Element.prototype.hasPointerCapture = () => false;
-      Element.prototype.setPointerCapture = () => {};
-      Element.prototype.releasePointerCapture = () => {};
-    }
-    if (!Element.prototype.scrollIntoView) {
-      Element.prototype.scrollIntoView = () => {};
-    }
   });
 
   afterEach(() => {
