+++ src/components/EmbeddedServerSidebar/EmbeddedServerItem.test.tsx
@@ -30,23 +30,6 @@ import { EmbeddedServerItem } from "./EmbeddedServerItem";
 import { TooltipProvider } from "@/components/ui";
 import { EmbeddedServerConfig, ServerState } from "@/types/embeddedServer";
 
-// jsdom lacks the observers/pointer-capture APIs Radix Tooltip touches when it
-// measures its content; shim them so the item's tooltip triggers can mount.
-class ResizeObserverStub {
-  observe() {}
-  unobserve() {}
-  disconnect() {}
-}
-if (!("ResizeObserver" in globalThis)) {
-  (globalThis as unknown as { ResizeObserver: typeof ResizeObserverStub }).ResizeObserver =
-    ResizeObserverStub;
-}
-if (!Element.prototype.hasPointerCapture) {
-  Element.prototype.hasPointerCapture = () => false;
-  Element.prototype.setPointerCapture = () => {};
-  Element.prototype.releasePointerCapture = () => {};
-}
-
 let container: HTMLDivElement;
 let root: Root;
