+++ src/components/EmbeddedServerSidebar/EmbeddedServerItem.tooltip.test.tsx
@@ -28,23 +28,6 @@ import { EmbeddedServerItem } from "./EmbeddedServerItem";
 import { TooltipProvider } from "@/components/ui";
 import { EmbeddedServerConfig, ServerState } from "@/types/embeddedServer";
 
-// jsdom lacks the observers/pointer-capture APIs Radix Tooltip touches when it
-// measures and portals its content; shim them so the primitive can mount.
-class ResizeObserverStub {
-  observe() {}
-  unobserve() {}
-  disconnect() {}
-}
-if (!("ResizeObserver" in globalThis)) {
-  (globalThis as unknown as { ResizeObserver: typeof ResizeObserverStub }).ResizeObserver =
-    ResizeObserverStub;
-}
-if (!Element.prototype.hasPointerCapture) {
-  Element.prototype.hasPointerCapture = () => false;
-  Element.prototype.setPointerCapture = () => {};
-  Element.prototype.releasePointerCapture = () => {};
-}
-
 let container: HTMLDivElement;
 let root: Root;
