+++ src/components/NetworkTools/HttpMonitorPanel.start-guard.test.tsx
@@ -15,33 +15,11 @@
  * exactly ONE start call.
  */
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import React, { act } from "react";
+import { act } from "react";
 import { createRoot, Root } from "react-dom/client";
 import { networkHttpMonitorStart } from "@/services/networkApi";
 import { HttpMonitorPanel } from "./HttpMonitorPanel";
-import { TooltipProvider } from "@/components/ui";
-
-// jsdom lacks the observer/pointer-capture APIs Radix Tooltip touches when it
-// mounts the tooltip-wrapped icon buttons; shim them so the panel renders.
-class ResizeObserverStub {
-  observe() {}
-  unobserve() {}
-  disconnect() {}
-}
-if (!("ResizeObserver" in globalThis)) {
-  (globalThis as unknown as { ResizeObserver: typeof ResizeObserverStub }).ResizeObserver =
-    ResizeObserverStub;
-}
-if (!Element.prototype.hasPointerCapture) {
-  Element.prototype.hasPointerCapture = () => false;
-  Element.prototype.setPointerCapture = () => {};
-  Element.prototype.releasePointerCapture = () => {};
-}
-
-/** Wrap a subtree so the shared Tooltip primitive finds its provider. */
-function withTooltip(ui: React.ReactElement): React.ReactElement {
-  return <TooltipProvider delayDuration={0}>{ui}</TooltipProvider>;
-}
+import { withTooltip } from "@/test/tooltip";
 
 // A deferred start: the promise stays pending until we resolve it, letting us
 // fire a second click while the first start is genuinely in flight.
