+++ src/components/NetworkTools/NetworkToolsSidebar.live.test.tsx
@@ -11,35 +11,13 @@
  * up reactively (its first check fires immediately on start).
  */
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import React, { act } from "react";
+import { act } from "react";
 import { createRoot, Root } from "react-dom/client";
 import { networkHttpMonitorList, onHttpMonitorCheck } from "@/services/networkApi";
 import type { HttpMonitorState, HttpCheckResult } from "@/types/network";
 import { useAppStore } from "@/store/appStore";
 import { NetworkToolsSidebar } from "./NetworkToolsSidebar";
-import { TooltipProvider } from "@/components/ui";
-
-// jsdom lacks the observer/pointer-capture APIs Radix Tooltip touches when it
-// mounts the tooltip-wrapped icon buttons; shim them so the sidebar renders.
-class ResizeObserverStub {
-  observe() {}
-  unobserve() {}
-  disconnect() {}
-}
-if (!("ResizeObserver" in globalThis)) {
-  (globalThis as unknown as { ResizeObserver: typeof ResizeObserverStub }).ResizeObserver =
-    ResizeObserverStub;
-}
-if (!Element.prototype.hasPointerCapture) {
-  Element.prototype.hasPointerCapture = () => false;
-  Element.prototype.setPointerCapture = () => {};
-  Element.prototype.releasePointerCapture = () => {};
-}
-
-/** Wrap a subtree so the shared Tooltip primitive finds its provider. */
-function withTooltip(ui: React.ReactElement): React.ReactElement {
-  return <TooltipProvider delayDuration={0}>{ui}</TooltipProvider>;
-}
+import { withTooltip } from "@/test/tooltip";
 
 vi.mock("@/services/networkApi", () => ({
   networkHttpMonitorList: vi.fn(() => Promise.resolve([])),
