+++ src/components/NetworkTools/WolPanel.test.tsx
@@ -6,32 +6,16 @@
  * (toast on success/failure) per the reactive design rule.
  */
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import React, { act } from "react";
+import { act } from "react";
 import { createRoot, Root } from "react-dom/client";
 import {
   networkWolSend,
   networkWolDevicesList,
   networkWolDeviceDelete,
 } from "@/services/networkApi";
-import { toast, TooltipProvider } from "@/components/ui";
+import { toast } from "@/components/ui";
+import { withTooltip } from "@/test/tooltip";
 import { WolPanel } from "./WolPanel";
-
-// jsdom lacks the observer/pointer-capture APIs Radix Tooltip touches when it
-// mounts the tooltip-wrapped icon buttons; shim them so the panel renders.
-class ResizeObserverStub {
-  observe() {}
-  unobserve() {}
-  disconnect() {}
-}
-if (!("ResizeObserver" in globalThis)) {
-  (globalThis as unknown as { ResizeObserver: typeof ResizeObserverStub }).ResizeObserver =
-    ResizeObserverStub;
-}
-if (!Element.prototype.hasPointerCapture) {
-  Element.prototype.hasPointerCapture = () => false;
-  Element.prototype.setPointerCapture = () => {};
-  Element.prototype.releasePointerCapture = () => {};
-}
 
 vi.mock("@/services/networkApi", () => ({
   networkWolSend: vi.fn(() => Promise.resolve()),
@@ -57,11 +41,6 @@ async function flush() {
   await act(async () => {
     await Promise.resolve();
   });
-}
-
-/** Wrap so the shared Tooltip primitive finds its provider (zero delay). */
-function withTooltip(ui: React.ReactElement): React.ReactElement {
-  return <TooltipProvider delayDuration={0}>{ui}</TooltipProvider>;
 }
 
 async function renderPanel() {
