+++ src/components/ui/Tooltip.test.tsx
@@ -3,23 +3,6 @@ import React, { act } from "react";
 import { createRoot, Root } from "react-dom/client";
 import { Tooltip, TooltipProvider } from "./Tooltip";
 
-// jsdom lacks the observers/pointer-capture APIs Radix Tooltip touches when it
-// measures and portals its content; shim them so the primitive can mount.
-class ResizeObserverStub {
-  observe() {}
-  unobserve() {}
-  disconnect() {}
-}
-if (!("ResizeObserver" in globalThis)) {
-  (globalThis as unknown as { ResizeObserver: typeof ResizeObserverStub }).ResizeObserver =
-    ResizeObserverStub;
-}
-if (!Element.prototype.hasPointerCapture) {
-  Element.prototype.hasPointerCapture = () => false;
-  Element.prototype.setPointerCapture = () => {};
-  Element.prototype.releasePointerCapture = () => {};
-}
-
 let container: HTMLDivElement;
 let root: Root;
